Extract: rename UI files edit/read
